requirements.txt
Flask==1.1.2
numpy==1.20.2
pandas==1.2.3
tensorflow==2.10.0
gdown==4.5.3
uuid==1.30
matplotlib==3.6.2
Pillow==9.3.0
gunicorn==20.1.0
wheel==0.38.4
